ESV verse label change & renamed file
