+++ src/components/highlight/Highlights.js
@@ -10,7 +10,7 @@ import DeleteIcon from "@material-ui/icons/Delete";
 import { useFirebase } from "react-redux-firebase";
 import { isLoaded, isEmpty, useFirebaseConnect } from "react-redux-firebase";
 import { useSelector } from "react-redux";
-import { getBookbyCode, capitalize } from "../common/utillity";
+import { getBookbyCode, capitalize } from "../common/utility";
 import Close from "../common/Close";
 import Box from "@material-ui/core/Box";
